+++ modules/ismp/pallets/pallet/src/child_trie.rs
@@ -1,12 +1,29 @@
-//! pallet Ismp child trie
+// Copyright (C) 2023 Polytope Labs.
+// SPDX-License-Identifier: Apache-2.0
 
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// 	http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//! Child trie storage layout for pallet-ismp
+//!
+//! pallet-ismp leverages a child trie to store outgoing/incoming requests and responses. This is
+//! because child tries provide cheaper state proofs compared the global state trie. This module
+//! describes the storage layout in the child trie.
+
+use crate::{dispatcher::LeafMetadata, utils::ResponseReceipt, Config};
 use alloc::vec::Vec;
 use core::marker::PhantomData;
-
 use frame_support::storage::child;
-pub use sp_core::{storage::ChildInfo, H256};
-
-use crate::{dispatcher::LeafMetadata, primitives::ResponseReceipt, Config};
+use sp_core::{storage::ChildInfo, H256};
 
 /// Commitments for outgoing requests
 /// The key is the request commitment
@@ -24,33 +41,31 @@ pub struct ResponseCommitments<T: Config>(PhantomData<T>);
 /// The key is the request commitment
 pub struct ResponseReceipts<T: Config>(PhantomData<T>);
 
-/// Constant pallet ismp name for all substrate chains
-pub const PALLET_NAME: &'static str = "Ismp";
 /// Child trie prefix for all substrate chains
 pub const CHILD_TRIE_PREFIX: &'static [u8] = b"ISMP";
 
-/// Storage key for a request commitment
+/// Returns the storage key for a request commitment in the child trie
 pub fn request_commitment_storage_key(key: H256) -> Vec<u8> {
     let mut full_key = "RequestCommitments".as_bytes().to_vec();
     full_key.extend_from_slice(&key.0);
     full_key
 }
 
-/// Storage key for a response commitment
+/// Returns the storage key for a response commitment in the child trie
 pub fn response_commitment_storage_key(key: H256) -> Vec<u8> {
     let mut full_key = "ResponseCommitments".as_bytes().to_vec();
     full_key.extend_from_slice(&key.0);
     full_key
 }
 
-/// Storage key for a request receipt
+/// Returns the storage key for a request receipt in the child trie
 pub fn request_receipt_storage_key(key: H256) -> Vec<u8> {
     let mut full_key = "RequestReceipts".as_bytes().to_vec();
     full_key.extend_from_slice(&key.0);
     full_key
 }
 
-/// Storage key for a response receipt
+/// Returns the storage key for a response receipt in the child trie
 /// The request commitment is the key
 pub fn response_receipt_storage_key(key: H256) -> Vec<u8> {
     let mut full_key = "ResponseReceipts".as_bytes().to_vec();
