+++ modules/ismp/pallets/pallet/src/events.rs
@@ -12,7 +12,8 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-//! Core ISMP events
+
+//! Pallet event definitions
 
 use crate::{errors::HandlingError, Config, Event as PalletEvent, Pallet};
 use alloc::vec::Vec;
