+++ modules/ismp/pallets/pallet/src/impls.rs
@@ -1,10 +1,26 @@
+// Copyright (C) 2023 Polytope Labs.
+// SPDX-License-Identifier: Apache-2.0
+
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// 	http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//! Pallet methods
+
 use crate::{
     child_trie::{RequestCommitments, ResponseCommitments},
     dispatcher::{FeeMetadata, LeafMetadata},
     host::Host,
-    mmr::{Leaf, ProofKeys},
-    primitives::{LeafIndexAndPos, Proof},
-    weight_info::get_weight,
+    mmr::{Leaf, LeafIndexAndPos, Proof, ProofKeys},
+    weights::get_weight,
     ChallengePeriod, Config, ConsensusClientUpdateTime, ConsensusStates, Error, Event,
     LatestStateMachineHeight, Pallet, Responded,
 };
@@ -59,7 +75,7 @@ impl<T: Config> Pallet<T> {
             items: proof.items,
         };
 
-        Ok((leaves.into_iter().map(|leaf| leaf.into()).collect(), proof))
+        Ok((leaves, proof))
     }
 
     /// Provides a way to handle messages.
