+++ parachain/runtimes/gargantua/src/ismp.rs
@@ -39,7 +39,7 @@ use sp_runtime::Percent;
 
 use ismp::router::Timeout;
 use ismp_sync_committee::constants::sepolia::Sepolia;
-use pallet_ismp::{dispatcher::FeeMetadata, host::Host, primitives::ModuleId};
+use pallet_ismp::{dispatcher::FeeMetadata, host::Host, utils::ModuleId};
 use sp_std::prelude::*;
 use staging_xcm::latest::MultiLocation;
