+++ parachain/runtimes/gargantua/src/lib.rs
@@ -70,7 +70,7 @@ use frame_system::{
     EnsureRoot, Phase,
 };
 
-use pallet_ismp::primitives::Proof;
+use pallet_ismp::utils::Proof;
 pub use sp_consensus_aura::sr25519::AuthorityId as AuraId;
 use sp_mmr_primitives::{LeafIndex, INDEXING_PREFIX};
 pub use sp_runtime::{MultiAddress, Perbill, Permill};
